requirements.txt
streamlit         # For the UI
requests          # For making HTTP requests to the AI API
python-dotenv     # For loading API keys from .env
